Changed use of HttpUrlConnection.setFixedLengthStreamingMode() to long instead of int

+++ src/main/java/uk/ac/ucl/cs/cmic/giftcloud/restserver/HttpConnection.java
@@ -48,7 +48,7 @@ public interface HttpConnection {
 
     void setRequestProperty(String key, String value);
 
-    void setFixedLengthStreamingMode(int contentLength);
+    void setFixedLengthStreamingMode(long contentLength);
 
     void connect() throws IOException;
